Moved AysncPropXXXX test classes under org.scalatest to propspect-test, even though they are commentted out.
